feat: Supabase Util functions
